BetterProtoPNet/requirements.txt
matplotlib
numpy
opencv-python
torch
torchvision
Pillow
augmentor
